backend/services/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
bcrypt
pyjwt
pydantic-settings
google-auth
